Remove pywin32 requirement:
It was being used to get shell paths, and for a small UAC function.

If pywin32 was not installed, these paths defaulted to the current working
directory.  So to display the error, an unneeded directory was created in the
cwd, to install the translation (which normally would be in LocalAppData), to
display the message.  Making a failsafe that could actually locate LocalAppData
involved doing exactly what pywin32 was doing anyway, so there's no longer a
need for it.

After that, might as well do the UAC function through ctypes as well.